+++ packages/kit/src/di/index.ts
@@ -3,5 +3,4 @@ import "./tsyringe-patch";
 export * from "./keys";
 export * from "./decorators";
 export * from "./context";
-export * from "./middlewares";
 export * from "./container";