+++ packages/server/src/middlewares/index.ts
@@ -1,3 +1,2 @@
-export * from "./coors";
 export * from "./PassthroughMiddleware";
 export * from "./LogRequestMiddleware";