+++ packages/server/src/testing/index.ts
@@ -1,3 +1,4 @@
 export { testContext, runInTestContext } from "./context";
+export * from "./mocks";
 export * from "./pipeline";
 export * from "./server";