+++ packages/server/src/testing/pipeline.ts
@@ -1,6 +1,10 @@
+import type http from "http";
+import { randomUUID } from "crypto";
+import { useContextProperty } from "../context";
+import { PassthroughMiddleware } from "../middlewares";
 import { HttpCallMetadata, httpCall, httpPipeline } from "../pipeline";
 import { HttpCallPipelineDefinition, HttpCServerMiddleware, HttpCallAccess } from "../server";
-import { runInTestContext } from "./context";
+import { runInTestContext, testContext } from "./context";
 
 
 export function httpCallTester<T extends HttpCallPipelineDefinition<any>>(call: T): T["execute"];
@@ -49,4 +53,81 @@ export function httpPipelineTester<T extends HttpCallPipelineDefinition<any>>(mi
             path: "/",
             params: args
         }))) as any;
+}
+
+
+
+export type CallBuilderOptions = {
+    middlewares?: HttpCServerMiddleware[]
+}
+
+export class CallBuilder {
+    protected _headers = {};
+    protected _url: string | undefined;
+    protected _context = {};
+    protected _middlewares: HttpCServerMiddleware[];
+
+    constructor(options?: CallBuilderOptions) {
+        this._middlewares = options?.middlewares || [];
+    }
+
+    withHeaders(headers: Record<string, string | number>): this {
+        this._headers = { ...this._headers, ...headers };
+        return this;
+    }
+
+    withUrl(url: string): this {
+        this._url = url;
+        return this;
+    }
+
+    withContext(context: Record<string, any>): this {
+        this._context = { ...this._context, ...context };
+        return this;
+    }
+
+    withMiddlewares(middlewares: HttpCServerMiddleware[], mode: "join" | "set" = "join"): this {
+        if (mode === "set") {
+            this._middlewares = middlewares;
+        } else {
+            this._middlewares = [...this._middlewares || [], ...middlewares];
+        }
+        return this;
+    }
+
+    create<T extends () => (any | Promise<any>)>(...pipeline: [...(HttpCServerMiddleware | HttpCallMetadata)[], T]): (...args: Parameters<T>) => Promise<Awaited<ReturnType<T>>> {
+        const call = httpCall(...pipeline as any);
+        return httpPipelineTester([
+            PassthroughMiddleware(() => {
+                testContext.clear();
+                useContextProperty("requestId", randomUUID());
+                useContextProperty("request", this._buildRequest());
+                Object.entries(this._context).forEach(([key, value]) => {
+                    useContextProperty(key, value);
+                });
+            }),
+            ...this._middlewares
+        ], call);
+    }
+
+    async run<T extends () => any>(handler: () => any): Promise<Awaited<ReturnType<T>>>;
+    async run(): Promise<void>;
+    async run<T extends () => any>(handler?: () => any): Promise<Awaited<ReturnType<T>>> {
+        return await this.create(handler || (() => { }))();
+    }
+
+    private _buildRequest(): http.IncomingMessage {
+        return {
+            headers: normalizeHeaders(this._headers) as any,
+            url: this._url,
+        } as http.IncomingMessage;
+    }
+}
+
+
+function normalizeHeaders(headers?: Record<string, number>) {
+    if (!headers) return {};
+    return Object.fromEntries(
+        Object.entries(headers || {}).map(([key, value]) => [key.toLowerCase(), value])
+    );
 }